pyproject.toml
[project]
name = "fine-pixtral"
version = "0.1.0"
description = "Fine-tuning dataset generation for Pixtral vision model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.0.0",
    "pillow>=10.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "trl>=0.7.0",
    "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git",
    "mlflow>=3.1.4",
    "boto3>=1.26.0",
    "pydantic-xml>=2.17.3",
    "pydantic>=2.11.7",
    "pdf2image>=1.17.0",
    "click>=8.2.1",
]
